Update tearDown for unittests to disregard `.gitignore` files

+++ common/document_parser/lib/responsibility_parse/tests/unit/test_responsibility_parser.py
@@ -4,6 +4,7 @@ import sys
 import os
 from common.document_parser.lib.responsibility_parse import ResponsibilityParser
 import pandas as pd
+
 sys.path.append(os.path.dirname(__file__).replace("/tests/unit", ""))
 
 
@@ -22,7 +23,8 @@ class TestResponsibilityParser(unittest.TestCase):
     @classmethod
     def tearDown(self) -> None:
         for file in os.listdir(self.output_dir):
-            os.remove(os.path.join(self.output_dir, file))
+            if file != ".gitignore":
+                os.remove(os.path.join(self.output_dir, file))
 
     def test_extract_numbering(self):
         numbering_input_output_dict = {
@@ -44,7 +46,8 @@ class TestResponsibilityParser(unittest.TestCase):
             "(zz). Here is some text": ("(zz).", "Here is some text"),
         }
         for input, expected_output in numbering_input_output_dict.items():
-            self.assertEqual(self.responsibility_parser.extract_numbering(input), expected_output, msg=f"Error: test_extract_numbering failed on input: {input}")
+            self.assertEqual(self.responsibility_parser.extract_numbering(input), expected_output,
+                             msg=f"Error: test_extract_numbering failed on input: {input}")
 
     def test_parse_entities(self):
         if self.require_exact_entity_match:
@@ -56,7 +59,8 @@ class TestResponsibilityParser(unittest.TestCase):
             for input, expected_output in entities_input_output_dict.items():
                 actual_output = self.responsibility_parser.parse_entities(input)
                 actual_output.sort()
-                self.assertListEqual(actual_output, expected_output,msg=f"Error: test_parse_entities failed on input: {input}")
+                self.assertListEqual(actual_output, expected_output,
+                                     msg=f"Error: test_parse_entities failed on input: {input}")
         else:
             self.assertTrue(True)
 
@@ -123,10 +127,12 @@ class TestResponsibilityParser(unittest.TestCase):
             ## remove entities from check, since depending on what GraphRelations the person running the test uses this
             ## changes the results/unittest
             if not self.require_exact_entity_match:
-                [expected_output_dict.pop(entity_key) for expected_output_dict in input_output_dict['output'] for entity_key in self.entities_keys]
+                [expected_output_dict.pop(entity_key) for expected_output_dict in input_output_dict['output'] for
+                 entity_key in self.entities_keys]
                 [output_dict.pop(entity_key) for output_dict in actual_output for entity_key in self.entities_keys]
 
-            self.assertListEqual(actual_output, input_output_dict['output'],msg=f"Error: test_format_responsibility_results failed on test case: {test_description}")
+            self.assertListEqual(actual_output, input_output_dict['output'],
+                                 msg=f"Error: test_format_responsibility_results failed on test case: {test_description}")
 
     def test_split_text_with_role_midline(self):
         input_output_dict = {
@@ -148,33 +154,36 @@ class TestResponsibilityParser(unittest.TestCase):
         }
         for input, expected_output in input_output_dict.items():
             actual_output = self.responsibility_parser.split_text_with_role_midline(input)
-            self.assertTupleEqual(actual_output, expected_output,msg=f"Error: test_split_text_with_role_midline failed on input: {input}")
+            self.assertTupleEqual(actual_output, expected_output,
+                                  msg=f"Error: test_split_text_with_role_midline failed on input: {input}")
 
     def test_parse_responsibility_lines_dodi_5000_94(self):
-        with open(os.path.join(self.input_dir,"DoDI 5000.94_resp_section.txt"), "r") as file:
+        with open(os.path.join(self.input_dir, "DoDI 5000.94_resp_section.txt"), "r") as file:
             resp_section_text = "".join(file.readlines())
-        with open(os.path.join(self.input_dir,"DoDI 5000.94_resp_section_expected.txt"), "r") as file:
+        with open(os.path.join(self.input_dir, "DoDI 5000.94_resp_section_expected.txt"), "r") as file:
             expected_output = eval("".join(file.readlines()))
 
         actual_output = self.responsibility_parser.parse_responsibility_section(resp_section_text)
-        self.assertListEqual(actual_output,expected_output)
+        self.assertListEqual(actual_output, expected_output)
 
     def test_extract_responsibilities_from_json_dodi_5000_94(self):
         actual_output = self.responsibility_parser.extract_responsibilities_from_json(json_filepath=
-                                                                                      os.path.join(self.input_dir,"DoDI 5000.94.json"))
-        with open(os.path.join(self.input_dir,"DoDI 5000.94_expected.txt"), "r") as file:
+                                                                                      os.path.join(self.input_dir,
+                                                                                                   "DoDI 5000.94.json"))
+        with open(os.path.join(self.input_dir, "DoDI 5000.94_expected.txt"), "r") as file:
             expected_output = eval("".join(file.readlines()))
 
         if not self.require_exact_entity_match:
             [expected_output_dict.pop(entity_key) for expected_output_dict in expected_output for entity_key
              in self.entities_keys]
-            [actual_output_dict.pop(entity_key) for actual_output_dict in actual_output for entity_key in self.entities_keys]
+            [actual_output_dict.pop(entity_key) for actual_output_dict in actual_output for entity_key in
+             self.entities_keys]
 
-        self.assertListEqual(actual_output,expected_output)
+        self.assertListEqual(actual_output, expected_output)
 
     def test_main(self):
         actual_output_filepath = os.path.join(self.output_dir, "actual_responsibility_results.xlsx")
-        expected_output = pd.read_excel(os.path.join(self.input_dir,"expected_responsibility_results.xlsx"))
+        expected_output = pd.read_excel(os.path.join(self.input_dir, "expected_responsibility_results.xlsx"))
 
         self.responsibility_parser.main(self.input_dir,
                                         excel_save_filepath=actual_output_filepath)
@@ -182,18 +191,17 @@ class TestResponsibilityParser(unittest.TestCase):
         # the blank file should throw an error
         self.assertSetEqual(self.responsibility_parser.error_files, {"blank_file.json"})
         # the "file_missing_responsibilties.json" should be logged as a file that doesn't contain any responsibilities
-        self.assertSetEqual(self.responsibility_parser.files_missing_responsibility_section,{"file_missing_responsibilities.json"})
+        self.assertSetEqual(self.responsibility_parser.files_missing_responsibility_section,
+                            {"file_missing_responsibilities.json"})
         # Verify out file was created
         self.assertTrue(os.path.isfile(actual_output_filepath))
         actual_output = pd.read_excel(actual_output_filepath)
 
         if not self.require_exact_entity_match:
-            expected_output = expected_output.drop(columns = self.entities_keys)
+            expected_output = expected_output.drop(columns=self.entities_keys)
             actual_output = actual_output.drop(columns=self.entities_keys)
 
-        pd.testing.assert_frame_equal(expected_output,actual_output)
-
-
+        pd.testing.assert_frame_equal(expected_output, actual_output)
 
 
 if __name__ == "__main__":
